Changes function name and handles over_sampled=0.
Changes the name of tv_1d to total_variation.

Also makes a somewhat significant change to the code to calculate
the spectrogram. Now properly handles the when over_sampled=0. This
allows somebody to calculate the spectrogram and NOT cut out the
over-sampled frequencies, which is interesting if you want to see
the data at the edges of the bandpass filters.