Update on entropy calculation, docstring and readme.
After testing the code with various distributions, and
rethinking the initial derivation, the code was changed and
documentation accordingingly.

It's unclear how to interpret the resulting entropy calculations
using the automatic histograming techniques in numpy and astroml.
In particular, it may be that the calculation of the maximum
entropy for varying sizes of bins and varying number of bins
is incorrect.

Nevertheless, the entropy calculation does seem to be sensitive
to the presence of a signal, based on preliminary testing with
simulated signals.