Fixes calculation of frequency bin edges for spectrogram.
Previously, the center frequency of the spectrogram was calculated
incorrectly, assuming that it was equal to the 'rf_center_frequency'
of the header. It turns out that 'rf_center_frequency' is the
center frequency of the first subband in the data file. Therefore,
one needs to correct for this fact.